backend/passSlip/serializers.py
from rest_framework import serializers
from .models import User, Slip, Code, Admin

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = '__all__'
        

class SlipSerializer(serializers.ModelSerializer):
    class Meta:
        model = Slip
        fields = '__all__'

class CodeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Code
        fields = '__all__'

class AdminSerializer(serializers.ModelSerializer):
    class Meta:
        model = Admin
        fields = '__all__'