requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.1.1
django-cors-headers==4.4.0
django-sslserver==0.22
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==25.0
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.9.0
